Add permissions logic to image input
